fix: address review feedback for queue polling
- Remove immediate update() call; WebSocket status event handles initial sync
- Only poll when activeJobsCount === 1 (last job, most likely to miss)
- Reset timer after any update completes to avoid redundant requests
- Add soft exponential backoff (8s -> 12s -> 18s -> 27s -> 32s cap),
  resets when job count changes
- Move to src/platform/remote/comfyui/ per project structure conventions
